Added sort order, shortened regexs
